requirements.txt
matplotlib==3.6.2
numpy==1.23.5
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
progress==1.6
scikit_image==0.19.3
scipy==1.9.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
